backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
django-cors-headers>=4.3
pytest>=8.0
pytest-django>=4.8
pytest-cov